% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellhash_calls.R
\name{add_barcode_table}
\alias{add_barcode_table}
\title{Add cellhash sample information to altExperiment rowData}
\usage{
add_barcode_table(sce, barcode_table, altexp_id = "cellhash")
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{barcode_table}{A data_frame of barcode_id and sample_id.}

\item{altexp_id}{The name of the alternative experiment that contains the cellhash data.
Default value: "cellhash"}
}
\value{
SingleCellExperiment with colData slot containing the demultiplexing calls
}
\description{
Add cellhash sample information to altExperiment rowData
}
\examples{
\dontrun{
}
}
