% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cell_mito_qc.R
\name{add_cell_mito_qc}
\alias{add_cell_mito_qc}
\title{Calculate QC metrics for all reads and mitochondrial subset for each cell in a SingleCellExperiment}
\usage{
add_cell_mito_qc(sce, mito, miQC = FALSE, ...)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{mito}{Character vector of mitochondrial gene names in the same format as rownames of SingleCellExperiment object.}

\item{miQC}{Logical indicating whether or not to calculate the posterior probability of a cell being compromised using
the linear mixture model in miQC. Default is FALSE.}

\item{...}{Any additional arguments to be passed to scuttle::addPerCellQCMetrics.}
}
\value{
SingleCellExperiment with colData slot containing calculated QC metrics.
 Columns include sum, detected, mito_sum, mito_detected, mito_percent, and total.
}
\description{
Calculate QC metrics for all reads and mitochondrial subset for each cell in a SingleCellExperiment
}
\examples{
\dontrun{
# add per cell QC metrics using only genes detected > 5\% of cells
add_cell_mito_qc(
  sce = sce,
  mito = mito_genes,
  threshold = 5
)
}
}
