% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellhash_calls.R
\name{add_demux_hashedDrops}
\alias{add_demux_hashedDrops}
\title{Add cellhash demultiplexing results using DropletUtils::hashedDrops}
\usage{
add_demux_hashedDrops(sce, altexp_id = "cellhash", ...)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{altexp_id}{The name of the alternative experiment that contains the cellhash data.
Default value: "cellhash"}

\item{...}{Other arguments to pass to DropletUtils::hashedDrops}
}
\value{
SingleCellExperiment with a `colData` column `hashedDrops_sampleid` containing the confident demultiplexing calls.
  Other results from `DropletUtils::hashedDrops()` are included in the `rowData` for the cellhash `altExp`,
  with the prefix `hashedDrops_`. See [DropletUtils::hashedDrops()] for the contents of these fields.
}
\description{
Add cellhash demultiplexing results using DropletUtils::hashedDrops
}
\examples{
\dontrun{
# add cell calls from DropletUtils::hashedDrops()
add_demux_hashedDrops(sce = sce)
}
}
