% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellhash_calls.R
\name{add_demux_seurat}
\alias{add_demux_seurat}
\title{Add cellhash demultiplexing results using Seurat::HTODemux()}
\usage{
add_demux_seurat(sce, altexp_id = "cellhash", ...)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{altexp_id}{The name of the alternative experiment that contains the cellhash data.
Default value: "cellhash"}

\item{...}{Other arguments to pass to Seurat::HTODemux}
}
\value{
SingleCellExperiment with a `colData` column `htodemux_id` containing the confident demultiplexing calls.
  Other results from `Seurat::HTODemuxs()` are included in the `rowData` for the cellhash `altExp`,
  with the prefix `HTODemux_`. See [Seurat::HTODemux()] for the contents of these fields.
}
\description{
Add cellhash demultiplexing results using Seurat::HTODemux()
}
\examples{
\dontrun{
  # add cell calls from Seurat::HTODemux()
  add_demux_seurat(sce = sce)
}
}
