% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vireo_calls.R
\name{add_demux_vireo}
\alias{add_demux_vireo}
\title{Add genetic demultiplexing results to a SCE object from Vireo}
\usage{
add_demux_vireo(sce, vireo_table)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{vireo_table}{A data frame of vireo results.
The expected format is that of the vireo `donor_ids.tsv` file, minimally containing columns
'cell' and 'donor_id' (but likely to contain other statistics as well).}
}
\value{
SingleCellExperiment with a `colData` column `vireo_sampleid` containing the confident demultiplexing calls.
  Other results imported from the vireo data frame are included in `colData` as well,
  with the prefix `vireo_`. See https://vireosnp.readthedocs.io/ for more information.
}
\description{
Add genetic demultiplexing results to a SCE object from Vireo
}
\examples{
\dontrun{
# add cell calls from a vireo data frame
add_demux_vireo(sce, vireo_table = vireo_df)
}
}
