% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_symbols.R
\name{add_gene_symbols}
\alias{add_gene_symbols}
\title{Add Gene symbols to a SingleCellExperiment from a GTF file}
\usage{
add_gene_symbols(sce, gene_info)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{gene_info}{A GRanges object or data frame with annotation information corresponding to the SCE.
Most importantly should contain a `gene_id` column that matches the row names of the SCE
and a `gene_name` column with the gene symbols should be included.}
}
\value{
SingleCellExperiment with gene_symbols added (or replaced) to the rowData table.
}
\description{
Add Gene symbols to a SingleCellExperiment from a GTF file
}
