% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellhash_calls.R
\name{add_hashsample_table}
\alias{add_hashsample_table}
\title{Add cellhash sample information to altExperiment rowData}
\usage{
add_hashsample_table(sce, hashsample_table, altexp_id = "cellhash")
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{hashsample_table}{A data_frame of barcode_id and sample_id.}

\item{altexp_id}{The name of the alternative experiment that contains the cellhash data.
Default value: "cellhash"}
}
\value{
SingleCellExperiment with rowData for the altExp containing barcode and sample ids
}
\description{
Add cellhash sample information to altExperiment rowData
}
\examples{
\dontrun{
# add sample information for a cellhash table
add_hashsample_table(sce = sce,
                     hashsample_table = barcode_ids)
}
}
