% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sample_metadata.R
\name{add_sample_metadata}
\alias{add_sample_metadata}
\title{Add data frame with sample metadata to an SCE object}
\usage{
add_sample_metadata(sce, metadata_df)
}
\arguments{
\item{sce}{SingleCellExperiment object to add metadata too}

\item{metadata_df}{A dataframe with metadata to add to the sce.}
}
\value{
SingleCellExperiment object with a `metadata_df` added to the `sample_metadata`
  slot of the metadata list
}
\description{
Add data frame with sample metadata to an SCE object
}
