% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{build_na_matrix}
\alias{build_na_matrix}
\title{Create a sparse matrix with all NA values}
\usage{
build_na_matrix(assay_name, matrix_rownames, matrix_colnames)
}
\arguments{
\item{assay_name}{Intended assay name for this matrix (e.g., "counts")}

\item{matrix_rownames}{Vector of matrix rown ames}

\item{matrix_colnames}{Vector of matrix column names}
}
\value{
Sparse matrix
}
\description{
Create a sparse matrix with all NA values
}
