% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{build_new_altexp_assay}
\alias{build_new_altexp_assay}
\title{Build a new sparse matrix for merging altExps}
\usage{
build_new_altexp_assay(
  assay_name,
  altexp_list,
  merged_rownames,
  all_merged_barcodes
)
}
\arguments{
\item{assay_name}{Name of assay of interest (e.g., "counts")}

\item{altexp_list}{List of altExps which should be included in the new matrix}

\item{merged_row_names}{Vector of matrix row names, corresponding to the full
set of features for this altExp}

\item{merged_col_names}{Vector of matrix column names, corresponding to all cells
which will be in the final merged altExp}
}
\value{
Sparse matrix
}
\description{
Build a new sparse matrix for merging altExps
}
