% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{build_new_altexp_rowdata}
\alias{build_new_altexp_rowdata}
\title{Build a new rowData slot for the merged altExp}
\usage{
build_new_altexp_rowdata(
  rowdata_list,
  all_merged_features,
  preserve_rowdata_cols = c("target_type")
)
}
\arguments{
\item{rowdata_list}{Names list of current rowData slots, in `data.frame` format}

\item{all_merged_features}{Vector of the full set of features for this altExp}

\item{preserve_rowdata_cols}{Vector of columns whose names should not be changed
and whose values are expected to apply to all altExps.}
}
\value{
Updated rowData slot object in `DataFrame` format
}
\description{
Build a new rowData slot for the merged altExp
}
