% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ilisi.R
\name{calculate_ilisi}
\alias{calculate_ilisi}
\title{Calculate iLISI (integration Local Inverse Simpson's Index) scores}
\usage{
calculate_ilisi(integrated_sce, pc_name, batch_column = "library_id")
}
\arguments{
\item{integrated_sce}{The integrated SCE object}

\item{pc_name}{The name that allows access to the PCs. Example: fastMNN_PCA}

\item{batch_column}{The variable in `integrated_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}
}
\value{
Data frame with three columns, one row per cell. Columns are `ilisi_score`,
  `cell_barcode`, and `batch_id`
}
\description{
Calculate iLISI (integration Local Inverse Simpson's Index) scores
}
