% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_silhouette_width.R
\name{calculate_silhouette_width}
\alias{calculate_silhouette_width}
\title{Calculate silhouette width scores from an integrated SCE object}
\usage{
calculate_silhouette_width(
  integrated_sce,
  pc_name,
  frac_cells = 0.8,
  nreps = 20,
  seed = 2023,
  batch_column = "library_id"
)
}
\arguments{
\item{integrated_sce}{The integrated SCE object}

\item{pc_name}{The name that allows access to the PCs. Example: fastMNN_PCA}

\item{frac_cells}{The fraction of cells to downsample to. Default: 0.8}

\item{nreps}{The number of times to repeat sub-sampling procedure. Default: 20}

\item{seed}{Seed for initializing random sampling}

\item{batch_column}{The variable in `integrated_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}
}
\value{
Tibble with six columns: `rep`, representing the given downsampling replicate;
  `silhouette_width`, the calculated silhouette width for the given `rep`; `silhouette_cluster`,
  the assigned cluster for the cell during silhouette calculation, i.e. the true identity;
  `other_cluster`, the other assigned for the cell during silhouette calculation;
  `pc_name`, the name associated with the pc results
}
\description{
Calculate silhouette width scores from an integrated SCE object
}
