% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_pcs.R
\name{calculate_silhouette_width_pcs}
\alias{calculate_silhouette_width_pcs}
\title{Calculate silhouette width scores from a merged SCE object.}
\usage{
calculate_silhouette_width_pcs(
  merged_sce,
  pc_name,
  batch_column = "library_id"
)
}
\arguments{
\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{pc_name}{The name used to access the PCA results stored in the
SingleCellExperiment object}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}
}
\value{
Tibble with five columns: `rep`, representing the given downsampling replicate;
  `silhouette_width`, the calculated silhouette width for the given `rep`; `silhouette_cluster`,
  the assigned cluster for the cell during silhouette calculation, i.e. the true identity;
  `other_cluster`, the other assigned for the cell during silhouette calculation;
  `pc_name`, the name associated with the pc results
}
\description{
This function performs replicated calculations on downsampled data.
}
