% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{check_altexps}
\alias{check_altexps}
\title{Helper function to check altExp compatibility}
\usage{
check_altexps(sce_list)
}
\arguments{
\item{sce_list}{List of SCEs with altExps to check}
}
\value{
List of named lists with altExp information for use when preparing to merge,
  with each sublist formatted as:
  altexp_name = list(features = c(features), assays = c(assays))
}
\description{
Helper function to check altExp compatibility
}
