% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{check_metadata}
\alias{check_metadata}
\title{Helper function to check that expected metadata fields are present in a given
list of SCEs.}
\usage{
check_metadata(sce_list, expected_fields = c("library_id", "sample_id"))
}
\arguments{
\item{sce_list}{List of SCEs to check}

\item{expected_fields}{a vector of metadata fields that should be present}
}
\description{
The default expected fields are `library_id` and `sample_id`. If either is missing from
  any SCE, an error is thrown. This function does not return anything.
}
