% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Perform clustering on a SingleCellExperiment object}
\usage{
cluster_sce(sce, pca_column, cluster_type, seed = NULL, k = NULL, ...)
}
\arguments{
\item{sce}{SingleCellExperiment object that requires clustering}

\item{pca_column}{The SingleCellExperiment column name to reference the PCA results}

\item{cluster_type}{The type of clustering method to be performed - can be "kmeans", "walktrap", or "louvain"}

\item{seed}{Seed for reproducibility of clustering results}

\item{k}{The desired number of centers}
}
\value{
SingleCellExperiment object containing clustering results
}
\description{
Perform clustering on a SingleCellExperiment object
}
