% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_df.R
\name{coldata_to_df}
\alias{coldata_to_df}
\title{Convert colData from SingleCellExperiment to a data.frame}
\usage{
coldata_to_df(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object.}
}
\value{
data.frame containing colData from the SingleCellExperiment object
  and a column, "cell_id", corresponding to the cell_barcodes or columns of the
  SingleCellExperiment object.
}
\description{
Convert colData from SingleCellExperiment to a data.frame
}
