% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_intron_counts.R
\name{collapse_intron_counts}
\alias{collapse_intron_counts}
\title{Merge counts from intron reads with corresponding cDNA reads}
\usage{
collapse_intron_counts(
  counts,
  which_counts = c("cDNA", "intron"),
  intron_metadata_path
)
}
\arguments{
\item{counts}{Counts matrix in sparse matrix format.}

\item{which_counts}{The type of counts (cDNA or intron) to include if alignment to intronic regions is TRUE. Default is FALSE.}

\item{intron_metadata}{Full path to a two column tsv file containing gene names for both spliced and intronic regions.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Merge counts from intron reads with corresponding cDNA reads
}
\examples{
\dontrun{
collapse_intron_counts(counts, which_counts = "cDNA",
intron_metadata_path = "<path to intron metadata.tsv>")
}
}
