% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{create_merged_altexp}
\alias{create_merged_altexp}
\title{Prepare altExps for merge and create a list of merged altExps for each altExp name}
\usage{
create_merged_altexp(altexp_list, merged_colnames)
}
\arguments{
\item{altexp_list}{List of altexps to merge}

\item{merged_colnames}{Vector of column names (`{sce_name}-{barcode}`) to include
in the final merged altExp. This vector includes _all_ SCEs, not only those
with this altExp name.}
}
\value{
A list of merged altExps to include the final merged SCE object
}
\description{
Prepare altExps for merge and create a list of merged altExps for each altExp name
}
