% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{create_sce_with_all_features}
\alias{create_sce_with_all_features}
\title{Create a new SCE that contains all provided features, or reorder a provided
 SCE into the given feature order}
\usage{
create_sce_with_all_features(sce, all_features)
}
\arguments{
\item{sce}{SCE from which a new SCE will be created, or which will be reordered}

\item{all_features}{A vector of all features that need to be in the final SCE,
in this order}
}
\value{
SCE object with all features in the given order
}
\description{
Create a new SCE that contains all provided features, or reorder a provided
 SCE into the given feature order
}
