% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_for_metrics.R
\name{downsample_for_metrics}
\alias{downsample_for_metrics}
\title{Downsample PCs for use in integration metric calculations}
\usage{
downsample_for_metrics(integrated_pcs, pc_name, frac_cells)
}
\arguments{
\item{integrated_pcs}{The full set of integrated pcs}

\item{pc_name}{The name that allows access to the PCs. Example: fastMNN_PCA}

\item{frac_cells}{The fraction of cells to downsample to}
}
\value{
List with two items: `pcs`, the downsampled PCs; `batch_labels`, the 
 corresponding batch labels as integers for the downsampled PCs
}
\description{
Downsample PCs for use in integration metric calculations
}
