% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{extract_metadata_for_altexp}
\alias{extract_metadata_for_altexp}
\title{Helper function to extract main SCE metadata for inclusion in an altExp}
\usage{
extract_metadata_for_altexp(sce)
}
\arguments{
\item{sce}{SCE object to extract metadata from}
}
\value{
List with fields `library_id` and `sample_id`
}
\description{
Helper function to extract main SCE metadata for inclusion in an altExp
}
