% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter counts matrix using DropletUtils::emptyDrops}
\usage{
filter_counts(sce, fdr_cutoff = 0.01, seed = NULL, ...)
}
\arguments{
\item{sce}{SingleCellExperiment with unfiltered gene x cell counts matrix.}

\item{fdr_cutoff}{FDR cutoff to use for DropletUtils::emptyDrops.
Default is 0.01.}

\item{seed}{An optional random seed for reproducibility.}

\item{...}{Any arguments to be passed into DropletUtils::emptyDrops.}
}
\value{
SingleCellExperiment with filtered gene x cell matrix.
}
\description{
This function will filter a SingleCellExperiment object using DropletUtils::emptyDrops(),
  as well as any associated alternative experiments. If mean expression and percent detected
  were previously calculated in the columns `mean` and `detected`, respectively, these
  will be removed from both the main and alternative experiments.
}
\examples{
\dontrun{
filter_counts(sce = sce_object)
}
}
