% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_pcs.R
\name{filter_pcs}
\alias{filter_pcs}
\title{Remove cells with NA batch label from PCs and label rownames}
\usage{
filter_pcs(pcs, batches, rename_pcs = TRUE)
}
\arguments{
\item{pcs}{The PCs to be filtered and labeled}

\item{batches}{Vector of cell-wise batch information whose length is the number of
rows in `pcs`}

\item{rename_pcs}{Whether to rename the PC rownames with the cell's batch label.
Default is `TRUE`.}
}
\value{
PCs with NA batch cells removed and labeled rownames
}
\description{
Remove cells with NA batch label from PCs and label rownames
}
