% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qc_report.R
\name{generate_qc_report}
\alias{generate_qc_report}
\title{Generate a QC report from a SingleCellExperiment object}
\usage{
generate_qc_report(
  sample_name,
  unfiltered_sce,
  filtered_sce = NULL,
  output = NULL
)
}
\arguments{
\item{sample_name}{The name of the sample for report headers}

\item{unfiltered_sce}{A SingleCellExperiment object that the report will describe}

\item{filtered_sce}{An optional filtered single cell experiment derived from first}

\item{output}{The output file path that will be created.
If the file name does not include an extension, ".html" will be added automatically.
Any directories in the path will be created as needed.}
}
\value{
The full path of the output file
}
\description{
Generate a QC report from a SingleCellExperiment object
}
\examples{
\dontrun{
generate_qc_report("Sample 1", my_sce, output = "reports/sample1_report.html")
}

}
