% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{get_altexp_attributes}
\alias{get_altexp_attributes}
\title{Helper function to check altExp compatibility}
\usage{
get_altexp_attributes(sce_list)
}
\arguments{
\item{sce_list}{List of SCEs with altExps to check}
}
\value{
List of named lists with altExp information for use when preparing to merge,
  with each sublist formatted as:
  altexp_name = list(features = c(features), assays = c(assays), coldata = c(column, names))
  - features are the full set of altExp features
  - assays are the set of assays for the given altExp (e.g. counts and logcounts)
  - coldata are the colData columns associated with this altxp to retain in the main SCE
}
\description{
Helper function to check altExp compatibility
}
