% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_quant_data.R
\name{import_quant_data}
\alias{import_quant_data}
\title{Import Gene Expression Quantification Data for Single-Cell RNA-Seq}
\usage{
import_quant_data(
  quant_dir,
  tool = c("cellranger", "alevin", "alevin-fry", "kallisto"),
  intron_mode = FALSE,
  usa_mode = FALSE,
  which_counts = c("cDNA", "intron"),
  intron_metadata_path
)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{tool}{Type of tool used to create files (Alevin, Alevin-fry, Cellranger, or Kallisto).}

\item{intron_mode}{Boolean indicating if the files included alignment to intronic regions. Default is FALSE.}

\item{usa_mode}{Boolean indicating if Alevin-fry was used, if the USA mode was invoked. Default is FALSE.}

\item{which_counts}{The type of counts (cDNA or intron) to include if alignment to intronic regions is TRUE. Default is FALSE.}

\item{intron_metadata}{Full path to a two column tsv file containing gene names for both spliced and intronic regions.
Only required if intron_mode = TRUE and usa_mode = FALSE.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Imports the gene x cell matrix output from either Alevin, Alevin-fry, Cellranger, or Kallisto and returns a SingleCellExperiment.
}
\examples{
\dontrun{
import_quant_data(quant_dir,
tool = "cellranger")

import_quant_data(quant_dir,
tool = "alevin-fry",
intron_mode = FALSE,
usa_mode = TRUE)
}

}
