% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_fastmnn}
\alias{integrate_fastmnn}
\title{Integrate a list of SingleCellExperiment (SCE) objects using the fastMNN
approach from the `batchelor` package.
Source: http://www.bioconductor.org/packages/release/bioc/html/batchelor.html}
\usage{
integrate_fastmnn(combined_sce, batch_column, ...)
}
\arguments{
\item{combined_sce}{A combined SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{batch_column}{The variable in `combined_sce` indicating batches.}

\item{...}{Additional arguments to pass into `batchelor::fastMNN()`}
}
\value{
The integrated SCE object
}
\description{
Integrate a list of SingleCellExperiment (SCE) objects using the fastMNN
approach from the `batchelor` package.
Source: http://www.bioconductor.org/packages/release/bioc/html/batchelor.html
}
