% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_harmony}
\alias{integrate_harmony}
\title{Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html}
\usage{
integrate_harmony(
  combined_sce,
  batch_column,
  harmony_do_PCA,
  harmony_covariate_cols,
  ...
)
}
\arguments{
\item{combined_sce}{A combined SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{batch_column}{The variable in `combined_sce` indicating batches.}

\item{harmony_do_PCA}{Whether integration should perform PCA (`TRUE`) and start with logcounts, or
use existing PCs to start (`FALSE`).}

\item{harmony_covariate_cols}{A vector of other columns in the combined SCE besides
`batch_column` should be considered as covariates during integration.}

\item{...}{Additional arguments to pass into `harmony::harmonyMatrix()`}
}
\value{
The integrated SCE object
}
\description{
Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html
}
