% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_altexp.R
\name{merge_altexp}
\alias{merge_altexp}
\title{Add an alternative experiment to a SingleCellExperiment by "left join"}
\usage{
merge_altexp(sce, alt_exp, alt_name)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{alt_exp}{a second SummarizedExperiment to add as an alternative experiment
Most likely, this will also be a SingleCellExperiment, but it does not have to be.
The AltExp in the output will be a SingleCellExperiment}

\item{alt_name}{The name to use for the alternative experiment slot}
}
\value{
A SingleCellExperiment with an AltExp slot containing another SingleCellExperiment.
}
\description{
This function takes SingleCellExperiment (SCE) and adds an Alternative Experiment (AltExp) to it.
If the two experiments do not have the same set of cells, this will use the first
SCE as the base set, adding columns (cells) to the AltExp so that the matrices match on that dimension.
Columns (cells) that are not found in the base experiment are discarded.
Note that this function is not related to merging multiple SCE objects.
}
