% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_to_coldata.R
\name{metadata_to_coldata}
\alias{metadata_to_coldata}
\title{Add sample metadata to colData of SCE object}
\usage{
metadata_to_coldata(sce, batch_column = "library_id")
}
\arguments{
\item{sce}{SingleCellExperiment object to add sample metadata to colData.
Must contain the `batch_column` as a column in `colData`.}

\item{batch_column}{A character value giving the colData column name
to differentiate originating SingleCellExperiment objects. Often these values
are unique library IDs. Default value is `library_id`.}
}
\value{
SingleCellExperiment object with the contents of `metadata(sce)$sample_metadata`
  added as new columns to the `colData(sce)`.
}
\description{
Add sample metadata to colData of SCE object
}
