% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_miQC.R
\name{miQC_nofilter}
\alias{miQC_nofilter}
\title{miQC_nofilter}
\usage{
miQC_nofilter(sce, model = NULL)
}
\arguments{
\item{sce}{(SingleCellExperiment) Input data object.}

\item{model}{(flexmix) Output of mixtureModel function, which should be
explicitly called first to ensure stability of model parameters.
Default = NULL.}
}
\value{
Returns a SingleCellExperiment object, the same as the input except
  with a new column in colData, prob_compromised.
}
\description{
Add statistics from miQC without filtering.
Derived from miQC::filterCells
https://github.com/greenelab/miQC/blob/main/R/filterCells.R
Copyright (c) 2021, Greene Laboratory
}
