% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{prepare_altexp_for_merge}
\alias{prepare_altexp_for_merge}
\title{Prepare an alternative experiment for merging}
\usage{
prepare_altexp_for_merge(
  sce,
  sce_name,
  altexp_name,
  expected_assays,
  expected_features,
  batch_column,
  cell_id_column,
  preserve_rowdata_cols = c("target_type")
)
}
\arguments{
\item{sce}{SCE with altExp to prepare}

\item{sce_name}{SCE name}

\item{altexp_name}{Name of altExp of interest}

\item{expected_assays}{Vector of assays that should be in each altExp}

\item{expected_features}{Vector of features that should be in each altExp}

\item{batch_column}{The name of the batch column which will be added to the
colData slot.}

\item{cell_id_column}{The name of the cell_id column which will be added to the
colData slot.}

\item{preserve_rowdata_cols}{altExp rowData columns which should not be renamed}
}
\value{
An SCE with an updated altEcp
}
\description{
If the given `altexp_name` name is missing from the SCE, a new one with all
  NA assays will be created.
}
