% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{prepare_merged_metadata}
\alias{prepare_merged_metadata}
\title{Prepare an updated list of metadata for the merged SCE}
\usage{
prepare_merged_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{List of metadata to update}
}
\value{
Updated metadata list to include in the merged SCE
}
\description{
The metadata will be reformatted to contain the following fields:
- `library_id`:       vector of all library ids in the merged object
- `sample_id`:        vector of all sample ids in the merged object
- `library_metadata`: pre-merge metadata list for each library, named by SCE name
- `sample_metadata`:  a data frame with all distinct rows from the pre-merge sample_metadata data frames,
  with all values coerced to character
}
