% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{prepare_sce_for_merge}
\alias{prepare_sce_for_merge}
\title{Helper function to prepare an SCE object for merging}
\usage{
prepare_sce_for_merge(
  sce,
  sce_name,
  batch_column,
  barcode_column,
  shared_features,
  retain_coldata_cols,
  preserve_rowdata_cols,
  expected_coldata_names
)
}
\arguments{
\item{sce}{The SCE object to be prepared}

\item{sce_name}{The name of the SCE object}

\item{batch_column}{The name of the batch column will which be added to the
colData slot}

\item{barcode_column}{The name of the column to be created (or not if it
already exists) in the colData slot to store cell barcodes}

\item{shared_features}{A vector of features (genes) that all SCEs to be merged
have in common}

\item{retain_coldata_cols}{A vector of columns to retain in the colData slot}

\item{preserve_rowdata_cols}{A vector of rowData columns which should not be
renamed}

\item{expected_coldata_names}{A vector of column names that are expected to be
present in the SCE. Any missing columns will be created with `NA` values.}
}
\value{
An updated SCE that is prepared for merging
}
\description{
Helper function to prepare an SCE object for merging
}
