% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_alevin}
\alias{read_alevin}
\title{Read in counts data processed with Alevin or Alevin-fry}
\usage{
read_alevin(
  quant_dir,
  mtx_format = FALSE,
  intron_mode = FALSE,
  usa_mode = FALSE,
  round_counts = TRUE,
  library_id = NULL,
  sample_id = NULL,
  tech_version = NULL
)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.}

\item{mtx_format}{Logical indicating if input data is in matrix market format.
Default is FALSE.}

\item{intron_mode}{Logical indicating if the files included alignment to intronic regions.
If TRUE, the main "counts" assay will contain counts aligned to introns and an additional "spliced"
assay will contain counts aligned to spliced cDNA only.
Default is FALSE.}

\item{usa_mode}{Logical indicating if Alevin-fry was used, if USA mode was invoked.
Implies the input data is in matrix market format.
Default is FALSE.}

\item{round_counts}{Logical indicating in the count matrix should be rounded to integers on import.
Only used if `usa_mode` is FALSE. Default is TRUE.}

\item{library_id}{Optional library identifier}

\item{sample_id}{Optional sample identifier.
If multiplexed samples are included in a library, this may be a vector.}

\item{tech_version}{Technology or kit used to process library (i.e. 10Xv3, 10Xv3.1).}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix.
}
\description{
Read in counts data processed with Alevin or Alevin-fry
}
\examples{
\dontrun{

# Import output files processed with either Alevin or Alevin-fry with alignment to
# cDNA only, including only spliced cDNA in final counts matrix
read_alevin(quant_dir)

# Import output files processed with either Alevin or Alevin-fry with alignment to
# cDNA + introns and including all unspliced cDNA in final counts matrix
read_alevin(quant_dir,
            intron_mode = TRUE)

# Import output files processed with alevin-fry USA mode
# including all unspliced cDNA in final counts matrix
read_alevin(quant_dir,
            usa_mode = TRUE,
            intron_mode = TRUE)

}
}
