% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_alevin_mtx}
\alias{read_alevin_mtx}
\title{Read in counts data processed with Alevin-fry in with mtx output.}
\usage{
read_alevin_mtx(quant_dir)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.}
}
\value{
unfiltered and uncollapsed gene x cell counts matrix
}
\description{
Read in counts data processed with Alevin-fry in with mtx output.
}
