% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellranger.R
\name{read_cellranger}
\alias{read_cellranger}
\title{Read in counts data processed with Cell Ranger}
\usage{
read_cellranger(cellranger_h5_file)
}
\arguments{
\item{cellranger_h5_file}{Path to H5AD file output from Cell Ranger.}
}
\value{
SingleCellExperiment of gene x cell counts matrix
}
\description{
Read in counts data processed with Cell Ranger
}
\examples{
\dontrun{

# Import data from cellranger output file
read_cellranger(cellranger_h5_file)
}

}
