% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kallisto.R
\name{read_kallisto}
\alias{read_kallisto}
\title{Read in counts data processed with Kallisto}
\usage{
read_kallisto(
  quant_dir,
  intron_mode = FALSE,
  which_counts = c("cDNA", "intron"),
  intron_metadata_path
)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{intron_mode}{Boolean indicating if the files included alignment to intronic regions. Default is FALSE.}

\item{which_counts}{The type of counts (cDNA or intron) to include if alignment to intronic regions is TRUE. Default is FALSE.}

\item{intron_metadata_path}{Full path to a two column tsv file containing gene names for both spliced and intronic regions.
Only required if intron_mode = TRUE and usa_mode = FALSE.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Read in counts data processed with Kallisto
}
\examples{
\dontrun{
read_kallisto(quant_dir,
intron_mode = TRUE,
which_counts = "intron",
intron_metadata_path = "<path to intron metadata.tsv>")
}
}
