% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_tximport}
\alias{read_tximport}
\title{Read in counts data processed with Alevin or alevin-fry in tximport-compatible formats}
\usage{
read_tximport(quant_dir)
}
\arguments{
\item{quant_dir}{Path to alevin output directory.}
}
\value{
unfiltered & uncollapsed gene x cell counts matrix
}
\description{
Read in counts data processed with Alevin or alevin-fry in tximport-compatible formats
}
