% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_usa_mode}
\alias{read_usa_mode}
\title{Read in counts data processed with Alevin-fry in USA mode}
\usage{
read_usa_mode(quant_dir)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.}
}
\value{
unfiltered and uncollapsed gene x cell counts matrix
}
\description{
Read in counts data processed with Alevin-fry in USA mode
}
