% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batch_nas_from_pcs.R
\name{remove_batch_nas_from_pcs}
\alias{remove_batch_nas_from_pcs}
\title{Remove NA batch cells from PCs for use in integration metric calculations}
\usage{
remove_batch_nas_from_pcs(integrated_pcs, batches)
}
\arguments{
\item{integrated_pcs}{The PCs with NA batch cells to be removed}

\item{batches}{Vector of cell-wise batch information whose length is the number of
rows in `integrated_pcs`}
}
\value{
List with two items: `pcs`: PCs with NA batch cells removed and
  rownames assigned as batch; `indices`: Which rows were _retained_
}
\description{
Remove NA batch cells from PCs for use in integration metric calculations
}
