% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_df.R
\name{rowdata_to_df}
\alias{rowdata_to_df}
\title{Convert rowData from SingleCellExperiment to a data.frame}
\usage{
rowdata_to_df(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object.}
}
\value{
data.frame containing rowData from the SingleCellExperiment object
  and a column, "gene_id", corresponding to the gene names or rows of the
  SingleCellExperiment object.
}
\description{
Convert rowData from SingleCellExperiment to a data.frame
}
