% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_anndata.R
\name{sce_to_anndata}
\alias{sce_to_anndata}
\title{Convert SingleCellExperiment objects to AnnData file stored as HDF5 file}
\usage{
sce_to_anndata(sce, anndata_file)
}
\arguments{
\item{sce}{SingleCellExperiment object to be converted to AnnData as an HDF5 file}

\item{anndata_file}{Path to output AnnData file. Must be an `.h5` or `.hdf5`}
}
\value{
SingleCellExperiment object equivalent to the AnnData object found in the
  HDF5 file produced.
}
\description{
Convert SingleCellExperiment objects to AnnData file stored as HDF5 file
}
\examples{
\dontrun{
sce_to_anndata(sce = sce_object,
               anndata_file = "test_anndata.h5")
}
}
