% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpcaTools-package.R
\docType{package}
\name{scpcaTools-package}
\alias{scpcaTools-package}
\title{scpcaTools: Useful tools for analysis of single-cell RNA seq counts data}
\description{
The scpcaTools package contains a set of tools for working with single-cell and single-nuclei RNA-seq counts data.
Mainly, this package can work with data that has been produced using Alevin, Alevin-Fry, Cellranger, or Kallisto.
}
