% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pc_rownames.R
\name{set_pc_rownames}
\alias{set_pc_rownames}
\title{Label rownames of provided PCs and remove NA batch cells from PCs}
\usage{
set_pc_rownames(integrated_pcs, batches)
}
\arguments{
\item{integrated_pcs}{The PCs with unlabeled rownames}

\item{batches}{Vector of cell-wise batch information whose length is the number of
rows in `integrated_pcs`}
}
\value{
PCs with labeled rownames and NA batch cells removed
}
\description{
Label rownames of provided PCs and remove NA batch cells from PCs
}
