% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_silhouette_width.R
\name{silhouette_width_from_pcs}
\alias{silhouette_width_from_pcs}
\title{Calculate silhouette width scores using PCs from a merged SCE object for use
in integration metric calculations}
\usage{
silhouette_width_from_pcs(
  merged_sce,
  pc_name,
  batch_column = "library_id",
  frac_cells,
  nreps,
  seed = NULL
)
}
\arguments{
\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{pc_name}{The name used to access the PCA results stored in the
SingleCellExperiment object}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}

\item{frac_cells}{The fraction of cells to downsample to.}

\item{nreps}{The number of times to repeat sub-sampling procedure.}

\item{seed}{Seed for initializing random sampling}
}
\value{
Tibble with five columns: `rep`, representing the given downsampling replicate;
  `silhouette_width`, the calculated silhouette width for the given `rep`; `silhouette_cluster`,
  the assigned cluster for the cell during silhouette calculation, i.e. the true identity;
  `other_cluster`, the other assigned for the cell during silhouette calculation;
  `pc_name`, the name associated with the pc results
}
\description{
Calculate silhouette width scores using PCs from a merged SCE object for use
in integration metric calculations
}
