% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sce.R
\name{sim_sce}
\alias{sim_sce}
\title{Create a random SingleCellExperiment object}
\usage{
sim_sce(n_genes = 200, n_cells = 100, n_empty = 1000, n_groups = 3)
}
\arguments{
\item{n_genes}{The number of genes to simulate}

\item{n_cells}{The number of cells to create.}

\item{n_empty}{How many "empty" droplets to create.
This will generally be much larger than the number of cells.}

\item{n_groups}{How many expression groups (cell clusters, roughly) to create.}
}
\value{
A SingleCellExperiment object.
}
\description{
Create a random SingleCellExperiment object
}
\examples{
\dontrun{
  sim_sce(n_genes = 100, n_cells = 100, n_empty = 2000, n_groups = 4)
}

}
