% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{update_sce_metadata}
\alias{update_sce_metadata}
\title{Helper function to update SCE metadata for merging}
\usage{
update_sce_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{Current SCE metadata before modification}
}
\value{
Updated metadata list to store in the SCE
}
\description{
Helper function to update SCE metadata for merging
}
