% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_within_batch_ari.R
\name{within_batch_ari_from_pcs}
\alias{within_batch_ari_from_pcs}
\title{Calculate within-batch ARI using provided PCs for use in integration metric calculations}
\usage{
within_batch_ari_from_pcs(
  individual_sce_list,
  merged_sce,
  pc_name,
  batch_column = "library_id",
  barcode_column = "cell_id",
  BLUSPARAM = bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard"),
  ...
)
}
\arguments{
\item{individual_sce_list}{A named list of individual SCE objects. It is
assumed these have a reduced dimension slot with principal components named "PCA".
Column names of each object should match the contents of the `barcode_column` for the `merged_sce.`}

\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{pc_name}{The name used to access the PCA results stored in the
SingleCellExperiment object}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}

\item{barcode_column}{The variable in `merged_sce` indicating the original cell barcode.
Default is "cell_id".}

\item{BLUSPARAM}{A BlusterParam object specifying the clustering algorithm to
use and any additional clustering options. Default is
`bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard")`}

\item{...}{Additional arguments to provide to `bluster::clusterRows()` within `cluster_sce()`}
}
\value{
Tibble with three columns: `ari`, representing the calculated ARI values;
  `batch_id`, the batch id associated with each `ari`; `pc_name`, the name
  associated with the pc results
}
\description{
Calculate within-batch ARI using provided PCs for use in integration metric calculations
}
