% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellhash_calls.R
\name{add_cellhash_ids}
\alias{add_cellhash_ids}
\title{Add cellhash sample information to altExperiment rowData}
\usage{
add_cellhash_ids(
  sce,
  hashsample_table,
  altexp_id = "cellhash",
  remove_unlabeled = FALSE,
  replace_rownames = FALSE
)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{hashsample_table}{A data_frame of barcode_id and sample_id.}

\item{altexp_id}{The name of the alternative experiment that contains the cellhash data.
Default value: "cellhash"}

\item{replace_rownames}{Replace rownames for the altExp with sample_ids.
Requires `remove_unlabeled` to be TRUE. Default `FALSE`.}

\item{removed_unlabeled}{Should data for barcodes with no corresponding sample be removed
from the altExp. Default `FALSE`.}
}
\value{
SingleCellExperiment with rowData for the altExp containing barcode and sample ids
}
\description{
Add cellhash sample information to altExperiment rowData
}
\examples{
\dontrun{
# add sample information for a cellhash table
add_cellhash_ids(sce = sce,
                 hashsample_table = barcode_ids)
}
}
