% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vireo_calls.R
\name{add_demux_vireo}
\alias{add_demux_vireo}
\title{Add genetic demultiplexing results to a SCE object from Vireo}
\usage{
add_demux_vireo(sce, vireo_df)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{vireo_df}{A data frame of vireo results.
The expected format is that of the vireo `donor_ids.tsv` file, minimally containing columns
'cell' and 'donor_id' (but likely to contain other statistics as well).}
}
\value{
SingleCellExperiment with a `colData` column `hashedDrops_sampleid` containing the confident demultiplexing calls.
  Other results from `DropletUtils::hashedDrops()` are included in the `rowData` for the cellhash `altExp`,
  with the prefix `hashedDrops_`. See [DropletUtils::hashedDrops()] for the contents of these fields.
}
\description{
Add genetic demultiplexing results to a SCE object from Vireo
}
\examples{
\dontrun{
  # add cell calls from DropletUtils::hashedDrops()
  add_demux_hashedDrops(sce = sce)
}
}
