% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_miQC.R
\name{add_miQC}
\alias{add_miQC}
\title{Add miQC model and probability compromised to a SingleCellExperiment object}
\usage{
add_miQC(sce, seed = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{seed}{An optional random seed for reproducibility.}
}
\value{
SingleCellExperiment with prob_compromised column added to colData
  If the model fits properly, there will also be a miQC_model slot added to the metadata.
}
\description{
Add miQC model and probability compromised to a SingleCellExperiment object
}
