% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sce.R
\name{build_sce}
\alias{build_sce}
\title{Build the SCE object from the counts matrix}
\usage{
build_sce(counts, intron_mode = FALSE, round_counts = TRUE)
}
\arguments{
\item{counts}{Counts matrix with rownames corresponding to gene names and colnames corresponding to cell barcodes.
Can be a counts matrix with intron counts specified by -I or an alevin-fry "USA" matrix,
with intron counts marked by "-U" and ambiguous counts "-A".}

\item{intron_mode}{Logical indicating if the files included alignment to intronic regions.
If TRUE, the main "counts" assay will contain counts aligned to introns and an additional "spliced"
assay will contain counts aligned to spliced cDNA only.
Default is FALSE.}

\item{round_counts}{Logical indicating in the count matrix should be rounded to integers on import.
Default is TRUE.}
}
\value{
SingleCellExperiment object containing either just a counts assay with spliced cDNA only if
  `intron_mode` is FALSE. If `intron_mode` is TRUE, the counts assay will contain both spliced and unspliced
   counts and the spliced assay will contain the counts for just the spliced cDNA.
}
\description{
Build the SCE object from the counts matrix
}
