% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_within_batch_ari.R
\name{calculate_within_batch_ari}
\alias{calculate_within_batch_ari}
\title{Calculate within-batch ARI scores from a merged SCE object.}
\usage{
calculate_within_batch_ari(
  individual_sce_list,
  pc_names,
  merged_sce,
  batch_column = "library_id",
  seed = NULL
)
}
\arguments{
\item{individual_sce_list}{A named list of individual SCE objects. It is
assumed these have a reduced dimension slot with principal components named "PCA".}

\item{pc_names}{A list of names that allow access to the PCs in the merged SCE
object. Example: c("PCA", "fastMNN_PCA"). A within-batch ARI will be returned for each `pc_name`.}

\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}

\item{seed}{Seed for initializing random sampling}
}
\value{
Tibble with three columns: `ari`, representing the calculated ARI values;
  `batch_id`, the batch id associated with each `ari`; `pc_name`, the name
  associated with the pc results
}
\description{
Calculate within-batch ARI scores from a merged SCE object.
}
