% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_within_batch_ari.R
\name{calculate_within_batch_ari}
\alias{calculate_within_batch_ari}
\title{Calculate within-batch ARI scores from a merged SCE object.}
\usage{
calculate_within_batch_ari(
  individual_sce_list,
  pc_names,
  merged_sce,
  batch_column = "library_id",
  barcode_column = "barcode",
  BLUSPARAM = bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard"),
  seed = NULL,
  ...
)
}
\arguments{
\item{individual_sce_list}{A named list of individual SCE objects. It is
assumed these have a reduced dimension slot with principal components named "PCA".
Column names of each object should match the contents of the `barcode_column` for the `merged_sce.`}

\item{pc_names}{A list of names of the PCs in the merged SCE `reducedDim` slot
object. Example: c("PCA", "fastMNN_PCA"). A within-batch ARI will be returned for each `pc_name`.}

\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}

\item{barcode_column}{The variable in `merged_sce` indicating the original cell barcode.
Default is "barcode".}

\item{BLUSPARAM}{A BlusterParam object specifying the clustering algorithm to
use and any additional clustering options. Default is
`bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard")`}

\item{seed}{Seed for initializing random sampling}

\item{...}{Additional arguments to provide to `bluster::clusterRows()` within `cluster_sce()`}
}
\value{
Tibble with three columns: `ari`, representing the calculated ARI values;
  `batch_id`, the batch id associated with each `ari`; `pc_name`, the name
  associated with the pc results
}
\description{
Calculate within-batch ARI scores from a merged SCE object.
}
