% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Perform clustering on a SingleCellExperiment object}
\usage{
cluster_sce(sce, seed = 2023, pca_column, k_range)
}
\arguments{
\item{sce}{SingleCellExperiment object that requires clustering}

\item{seed}{Seed for reproducibility of clustering results}

\item{pca_column}{The SingleCellExperiment column name to reference the PCA results}

\item{k_range}{Range of k values to use for k-means clustering. Default: `seq(5, 25, 5)`}
}
\value{
SingleCellExperiment object containing clustering results
}
\description{
Perform clustering on a SingleCellExperiment object
}
