% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Perform clustering on a SingleCellExperiment object}
\usage{
cluster_sce(sce, pca_name, BLUSPARAM, cluster_column_name, seed = NULL, ...)
}
\arguments{
\item{sce}{SingleCellExperiment object that requires clustering}

\item{pca_name}{The name used to access the PCA results}

\item{BLUSPARAM}{A BlusterParam object specifying the clustering algorithm to
use}

\item{cluster_column_name}{The name of the column to store the clustering
results in the SCE}

\item{seed}{Seed for reproducibility of clustering results}
}
\value{
SingleCellExperiment object containing clustering results
}
\description{
Perform clustering on a SingleCellExperiment object
}
