% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_intron_counts.R
\name{collapse_intron_counts}
\alias{collapse_intron_counts}
\title{Merge counts from intron reads with corresponding cDNA reads}
\usage{
collapse_intron_counts(counts, which_counts = c("spliced", "unspliced"))
}
\arguments{
\item{counts}{Counts matrix with rownames corresponding to gene names and colnames corresponding to cell barcodes.
Can be a counts matrix with intron counts specified by -I or an alevin-fry "USA" matrix,
with intron counts marked by "-U" and ambiguous counts "-A".}

\item{which_counts}{If intron_mode is TRUE, which type of counts should be included:
Only counts aligned to spliced cDNA ("spliced") or all spliced and unspliced cDNA ("unspliced").
Ambiguous counts in USA mode are always included.
Default is "spliced".}
}
\value{
unfiltered gene x cell counts matrix
}
\description{
Merge counts from intron reads with corresponding cDNA reads
}
\examples{
\dontrun{

# only keep counts from spliced cDNA in final counts matrix
collapse_intron_counts(counts,
                       which_counts = "spliced")
}

# include unspliced cDNA in final counts matrix
collapse_intron_counts(counts,
                       which_counts = "unspliced")


}
