% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_pcs.R
\name{downsample_pcs}
\alias{downsample_pcs}
\title{Downsample PCs for use in integration metric calculations}
\usage{
downsample_pcs(pcs, frac_cells)
}
\arguments{
\item{pcs}{The PCs to downsample}

\item{frac_cells}{The fraction of cells to downsample to}
}
\value{
List with two items: `pcs`, the downsampled PCs; `batch_labels`, the
 corresponding batch labels as integers for the downsampled PCs
}
\description{
Downsample PCs for use in integration metric calculations
}
