% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_pcs.R
\name{downsample_pcs}
\alias{downsample_pcs}
\title{Downsample PCs for use in integration metric calculations}
\usage{
downsample_pcs(pcs, frac_cells)
}
\arguments{
\item{pcs}{The PCs to downsample, these PCs should contain batch labels as rownames}

\item{frac_cells}{The fraction of cells to downsample to}
}
\value{
The downsampled PCs
}
\description{
Downsample PCs for use in integration metric calculations
}
