% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter counts matrix using DropletUtils::emptyDrops}
\usage{
filter_counts(sce, fdr_cutoff = 0.01, ...)
}
\arguments{
\item{sce}{SingleCellExperiment with unfiltered gene x cell counts matrix.}

\item{fdr_cutoff}{FDR cutoff to use for DropletUtils::emptyDrops.
Default is 0.01.}

\item{...}{Any arguments to be passed into DropletUtils::emptyDrops.}
}
\value{
SingleCellExperiment with filtered gene x cell matrix.
}
\description{
Filter counts matrix using DropletUtils::emptyDrops
}
\examples{
\dontrun{
filter_counts(sce = sce_object)
}
}
