% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qc_report.R
\name{generate_qc_report}
\alias{generate_qc_report}
\title{Generate a QC report from a SingleCellExperiment object}
\usage{
generate_qc_report(sce, sample_name, output = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object that the report will describe}

\item{sample_name}{The name of the sample}

\item{outfile}{The output file path that will be created.
If the file name does not include an extension, ".html" will be added automatically.
Any directories in the path will be created as needed.}
}
\value{
The path of the output file
}
\description{
Generate a QC report from a SingleCellExperiment object
}
\examples{
\dontrun{
generate_qc_report(my_sce, "Sample 1", output = "reports/sample1_report.html")
}
}
