% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_quant_data.R
\name{import_quant_data}
\alias{import_quant_data}
\title{Import Gene Expression Quantification Data for Single-Cell RNA-Seq}
\usage{
import_quant_data(
  quant_dir,
  tool = c("cellranger", "alevin", "alevin-fry", "kallisto"),
  which_counts = c("spliced", "unspliced"),
  intron_mode = FALSE,
  usa_mode = FALSE
)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{tool}{Type of tool used to create files (Alevin, Alevin-fry, Cellranger, or Kallisto).}

\item{which_counts}{If intron_mode is TRUE, which type of counts should be included,
only counts aligned to spliced cDNA ("spliced") or all spliced and unspliced cDNA ("unspliced").
Default is "spliced".}

\item{intron_mode}{Logical indicating if the files included alignment to intronic regions.
Default is FALSE.}

\item{usa_mode}{Logical indicating if Alevin-fry was used, if the USA mode was invoked.
Default is FALSE.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Imports the gene x cell matrix output from either Alevin, Alevin-fry, Cellranger, or Kallisto and returns a SingleCellExperiment.
}
\examples{
\dontrun{
# read in single cell RNA seq data processed using cellranger
import_quant_data(quant_dir,
tool = "cellranger")

# read in single-cell RNA seq data processed using alevin-fry in USA mode
# with alignment to cDNA only and including counts for spliced cDNA only
import_quant_data(quant_dir,
  tool = "alevin-fry",
  intron_mode = FALSE,
  usa_mode = TRUE)

# read in single-nuclei RNA-seq data processed using alevin-fry in
# USA mode with alignment to cDNA + introns and including counts for
# unspliced cDNA
import_quant_data(quant_dir,
  tool = "alevin-fry",
  usa_mode = TRUE,
  which_counts = "unspliced")

# read in single-nuclei RNA-seq data processed using kallisto with
alignment to cDNA + introns and including counts for unspliced cDNA
import_quant_data(quant_dir,
  tool = "kallisto",
  intron_mode = TRUE,
  which_counts = "unspliced")
}

}
