% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_quant_data.R
\name{import_quant_data}
\alias{import_quant_data}
\title{Import Gene Expression Quantification Data for Single-Cell RNA-Seq}
\usage{
import_quant_data(
  quant_dir,
  cellranger_h5_file = "",
  tool = c("cellranger", "alevin", "alevin-fry", "kallisto"),
  include_unspliced = TRUE,
  usa_mode = FALSE,
  filter = FALSE,
  fdr_cutoff = 0.01,
  tech_version = NULL,
  ...
)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.
Used when tool is alevin, alevin-fry, or kallisto.}

\item{cellranger_h5_file}{Optional path to H5 file output by Cell Ranger.
Used when tool is cellranger.}

\item{tool}{Type of tool used to create files (alevin, alevin-fry, cellranger, or kallisto).}

\item{include_unspliced}{Whether or not to include the unspliced reads in the counts matrix.
If TRUE, the main "counts" assay will contain unspliced reads and spliced reads and an additional "spliced"
assay will contain spliced reads only. If TRUE, requires that data has been aligned to a reference containing
spliced and unspliced reads.
Default is TRUE.}

\item{usa_mode}{Logical indicating if Alevin-fry was used, if the USA mode was invoked.
Default is FALSE.}

\item{filter}{Logical indicating whether or not to filter the counts matrix.
Filtering is performed using DropletUtils::emptyDrops and cannot be performed with Cell Ranger.}

\item{fdr_cutoff}{FDR cutoff to use for DropletUtils::emptyDrops.
Default is 0.01.}

\item{tech_version}{Technology or kit used to process library (i.e. 10Xv3, 10Xv3.1).}

\item{...}{Any arguments to be passed into DropletUtils::emptyDrops.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Imports the gene x cell matrix output from either Alevin, Alevin-fry, Cell Ranger, or Kallisto
  and returns a SingleCellExperiment.
}
\examples{
\dontrun{
# read in single cell RNA seq data processed using cellranger
import_quant_data(cellranger_h5_file,
  tool = "cellranger"
)

# read in single-cell RNA seq data processed using alevin-fry in USA mode
# with alignment to cDNA only and including counts for spliced cDNA only
import_quant_data(quant_dir,
  tool = "alevin-fry",
  include_unspliced = FALSE,
  usa_mode = TRUE
)

# read in single-nuclei RNA-seq data processed using alevin-fry in
# USA mode with alignment to cDNA + introns and including counts for
# unspliced cDNA and perform filtering
import_quant_data(quant_dir,
  tool = "alevin-fry",
  usa_mode = TRUE,
  filter = TRUE
)

# read in single-nuclei RNA-seq data processed using kallisto with
# alignment to cDNA + introns and including counts for unspliced cDNA
# and perform filtering
import_quant_data(quant_dir,
  tool = "kallisto",
  filter = TRUE
)
}

}
