% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_harmony}
\alias{integrate_harmony}
\title{Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html}
\usage{
integrate_harmony(merged_sce, batch_column, covariate_cols = c(), ...)
}
\arguments{
\item{merged_sce}{A merged SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{batch_column}{The column in the merged SCE object indicating batches
being integrated.}

\item{covariate_cols}{A vector of other columns to consider as
covariates during integration.}

\item{...}{Additional arguments to pass into `harmony::HarmonyMatrix()`}
}
\value{
Integrated PCs as calculated by `harmony`
}
\description{
Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html
}
