% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_harmony}
\alias{integrate_harmony}
\title{Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html}
\usage{
integrate_harmony(combined_sce, covariate_cols, ...)
}
\arguments{
\item{combined_sce}{A combined SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{covariate_cols}{A vector of other columns, including the batch, to
consider as covariates during integration.}

\item{...}{Additional arguments to pass into `harmony::harmonyMatrix()`}

\item{batch_column}{The variable in `combined_sce` indicating batches.}
}
\value{
The integrated SCE object
}
\description{
Integrate a list of SingleCellExperiment (SCE) objects using the harmony
approach from the `harmony` package.
Source: https://cran.r-project.org/web/packages/harmony/index.html
}
