% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_sces}
\alias{integrate_sces}
\title{Integrate a merged set of SingleCellExperiment objects using a specified
 integration method. The final SCE object will contain an additional reducedDim
 entitled `\{integration_method\}_PCA`. All original SCE assays are retained.
 `fastMNN` integration uses all default setting, but additional parameters can be supplied.
 `harmony` integration uses existing PCs, but additional parameters can be supplied.}
\usage{
integrate_sces(
  merged_sce,
  integration_method = c("fastMNN", "harmony"),
  batch_column = "sample",
  covariate_cols = c(),
  batch_lambda = 1,
  covariate_lambda = c(),
  return_corrected_expression = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{merged_sce}{A merged SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{integration_method}{The integration method to use. One of `fastMNN` or
`harmony` (case-sensitive)}

\item{batch_column}{The column in the merged SCE object indicating batches}

\item{covariate_cols}{A vector of additional columns in the merged SCE to
consider as covariates during integration. Currently, this is used only by
`harmony`.}

\item{batch_lambda}{The ridge regression penalty to use when correcting using the `batch_column`.
This is only used by `harmony`. Default is 1.}

\item{covariate_lambda}{A vector of ridge regression penalties to use with each covariate.
A lambda value must be provided for each covariate column.
This is only used by `harmony`.}

\item{return_corrected_expression}{A boolean indicating whether corrected expression
values determined by the given integration method should be included in the
integrated SCE object. Note that `harmony` does not calculate corrected expression,
so this argument is ignored for this integration method. Default is `FALSE`.}

\item{seed}{Random seed to set for integration. This is only set if the value is not `NULL`.}

\item{...}{Any additional parameters to be passed to the given integration method}
}
\value{
An updated merged SCE object containing a new reduced dimension named
  `\{integration_method\}_PCA` containing the corrected PCs. In the case of `fastMNN`
  integration, the SCE also includes a corrected expression assay (`fastMNN_corrected`)
}
\description{
Integrate a merged set of SingleCellExperiment objects using a specified
 integration method. The final SCE object will contain an additional reducedDim
 entitled `\{integration_method\}_PCA`. All original SCE assays are retained.
 `fastMNN` integration uses all default setting, but additional parameters can be supplied.
 `harmony` integration uses existing PCs, but additional parameters can be supplied.
}
