% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sces.R
\name{integrate_sces}
\alias{integrate_sces}
\title{Integrate a combined set of SingleCellExperiment objects using a specified
 integration method. The final SCE object will contain an additional reducedDim
 entitled `{integration_method}_PCA`.}
\usage{
integrate_sces(
  combined_sce,
  integration_method,
  batch_column = "sample",
  retain_assays = c("counts", "logcounts"),
  return_corrected_expression = TRUE,
  harmony_do_PCA = FALSE,
  harmony_covariate_cols = c(),
  seed = NULL,
  ...
)
}
\arguments{
\item{combined_sce}{A combined SCE object as prepared by `scpcaTools::merge_sce_list()`.}

\item{integration_method}{The integration method to use. One of `fastMNN` or
`harmony` (case-insensitive)}

\item{batch_column}{The column in the combined SCE object indicating batches
being integrated. Default is "sample".}

\item{retain_assays}{A character vector of uncorrected assays to retain in
the integrated SCE object. By default, both "counts" and "logcounts" are retained.}

\item{return_corrected_expression}{A boolean indicating whether corrected expression
values determined by the given integration method should be included in the
integrated SCE object. Note that `harmony` does not calculate corrected expression,
so this argument is ignored for this integration method. Default is `TRUE`.}

\item{harmony_do_PCA}{A `harmony`-specific argument (ignored otherwise) indicating whether
integration should perform PCA. If `FALSE` (default), `harmony` will use the existing PCs to perform
integration. If `TRUE`, `harmony` will will start from normalized expression when integrating and
re-calculate PCs.}

\item{harmony_covariate_cols}{A `harmony`-specific argument (ignored otherwise)
providing other columns in the combined SCE besides `batch_column` that `harmony`
should consider as covariates during integration.}

\item{seed}{Random seed to set for integration. This is only set if the value is not `NULL`.}

\item{...}{Any additional parameters to be passed to the given integration method}
}
\value{
An update combined SCE object containing
}
\description{
Integrate a combined set of SingleCellExperiment objects using a specified
 integration method. The final SCE object will contain an additional reducedDim
 entitled `{integration_method}_PCA`.
}
