% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{merge_sce_list}
\alias{merge_sce_list}
\title{Merge a list of SCEs as preparation for formal integration}
\usage{
merge_sce_list(
  sce_list = list(),
  batch_column = "library_id",
  retain_coldata_cols = c("sum", "barcode", "detected", "total", "subsets_mito_sum",
    "subsets_mito_detected", "subsets_mito_percent", "miQC_pass", "prob_compromised"),
  preserve_rowdata_cols = NULL,
  cell_id_column = "cell_id"
)
}
\arguments{
\item{sce_list}{A list of SingleCellExperiment objects. The list may optionally
be named with batch information. If no names are provided, names will be generated
based on the SCE's index.}

\item{batch_column}{A character value giving the resulting colData column name
to differentiate originating SingleCellExperiment objects. Often these values
are unique library IDs. Default value is `library_id`.}

\item{retain_coldata_cols}{A vector of colData columns which should be retained
in the the final merged SCE object.}

\item{preserve_rowdata_cols}{A vector of column names that appear in originating
SCE objects' rowData slots which should not be renamed with
the given SCE object name or index is name is not given. These are generally
columns which are not specific to the given library's preparation or statistics.
For example, such a vector might contain items like "Gene", "ensembl_ids", etc.}

\item{cell_id_column}{A character value giving the resulting colData column name
to hold unique cell IDs formatted as their original row name. Default
value is `cell_id`.}
}
\value{
A SingleCellExperiment object containing all SingleCellExperiment objects
  present in the inputted list
}
\description{
This function takes an optionally-named (if named, ideally by a form of
 library ID) list of SingleCellExperiment (SCE) objects and merges them into
 one SCE object. At least some genes must be present in all SCEs in order to
 merge them. Currently any present altExps are not retained.
}
\details{
Original SCE contents are modified or retained as follows:
 - The resulting colData slot will include a new column specified by
   `batch_column` (default "library_id") that either holds the originating SCE
   object's name (referred to as `sce_name` here), or if it is unnamed then
   its index in the provided `sce_list`.
 - The resulting colData slot will include another new column `cell_id_column`
   (default "cell_id") that will contain the SCE's original column names (i.e.
   original colData rownames). Often, but not always, this rowname holds a
   unique cell barcode.
 - The resulting colData rownames will be be prefixed with `{sce_name-}`.
 - The resulting rowData slot column names will be appended with the given
   SCE's name, as `{sce_name}-{column_name}` except for columns whose names
   are indicated to preserve with the `preserve_rowdata_cols` argument.
}
