% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_miQC.R
\name{miQC_nofilter}
\alias{miQC_nofilter}
\title{miQC_nofilter}
\usage{
miQC_nofilter(sce, model = NULL, verbose = TRUE)
}
\arguments{
\item{sce}{(SingleCellExperiment) Input data object.}

\item{model}{(flexmix) Output of mixtureModel function, which should be
explicitly called first to ensure stability of model parameters.
Default = NULL.}
}
\value{
Returns a SingleCellExperiment object, the same as the input except
  with a new column in colData, prob_compromised.
}
\description{
Add statistics from miQC without filtering.
Derived from miQC::filterCells
https://github.com/greenelab/miQC/blob/main/R/filterCells.R
Copyright (c) 2021, Greene Laboratory
}
\examples{
library(scRNAseq)
library(scater)
sce <- ZeiselBrainData()
mt_genes <- grepl("^mt-",  rownames(sce))
feature_ctrls <- list(mito = rownames(sce)[mt_genes])
sce <- addPerCellQC(sce, subsets = feature_ctrls)
model <- mixtureModel(sce)
sce <- miQC_nofilter(sce, model)

}
