% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{prepare_sce_for_merge}
\alias{prepare_sce_for_merge}
\title{Helper function to prepare an SCE object for merging}
\usage{
prepare_sce_for_merge(
  sce,
  sce_name,
  batch_column,
  cell_id_column,
  shared_features,
  retain_coldata_cols,
  preserve_rowdata_cols,
  is_altexp = FALSE
)
}
\arguments{
\item{sce}{The SCE object to be prepared}

\item{sce_name}{The name of the SCE object}

\item{batch_column}{The name of the batch column which will be added to the
colData slot}

\item{cell_id_column}{The name of the cell_id column which will be added to the
colData slot}

\item{shared_features}{A vector of features (genes) that all SCEs to be merged
have in common}

\item{retain_coldata_cols}{A vector of columns to retain in the colData slot.
If columns are missing from the data, they will be filled with `NA` values.
The exceptions to this are `barcode_column` and `batch_column` which will be
populated with respective information.}

\item{preserve_rowdata_cols}{A vector of rowData columns which should not be
renamed}

\item{is_altexp}{Boolean if we are preparing an altExp or not.
Default is `FALSE`. If FALSE, the SCE colnames will not be modified.}
}
\value{
An updated SCE that is prepared for merging
}
\description{
Helper function to prepare an SCE object for merging
}
