% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_alevin}
\alias{read_alevin}
\title{Read in counts data processed with Alevin or Alevin-fry}
\usage{
read_alevin(
  quant_dir,
  intron_mode = FALSE,
  usa_mode = FALSE,
  which_counts = c("spliced", "unspliced")
)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{intron_mode}{Boolean indicating if the files included alignment to intronic regions.
Default is FALSE.}

\item{usa_mode}{Boolean indicating if Alevin-fry was used, if the USA mode was invoked.
Default is FALSE.}

\item{which_counts}{If intron_mode is TRUE, which type of counts should be included,
only counts aligned to spliced cDNA ("spliced") or all spliced and unspliced cDNA ("unspliced").
Default is "spliced".}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix.
}
\description{
Read in counts data processed with Alevin or Alevin-fry
}
\examples{
\dontrun{
read_alevin(quant_dir,
                     intron_mode = TRUE,
                     usa_mode = TRUE,
                     which_counts = "unspliced")
}
}
