% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_alevin}
\alias{read_alevin}
\title{Read in counts data processed with Alevin or Alevin-fry}
\usage{
read_alevin(
  quant_dir,
  intron_mode = FALSE,
  usa_mode = FALSE,
  which_counts = c("cDNA", "intron"),
  intron_metadata_path
)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{intron_mode}{Boolean indicating if the files included alignment to intronic regions. Default is FALSE.}

\item{usa_mode}{Boolean indicating if Alevin-fry was used, if the USA mode was invoked. Default is FALSE.}

\item{intron_metadata_path}{Full path to a two column tsv file containing gene names for both spliced and intronic regions.
Only required if intron_mode = TRUE and usa_mode = FALSE.}

\item{tool}{Type of tool used to create files (Alevin or Alevin-fry).}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix.
}
\description{
Read in counts data processed with Alevin or Alevin-fry
}
\examples{
\dontrun{
read_alevin(quant_dir,
intron_mode = TRUE,
usa_mode = TRUE,
which_counts = "intron")
}
}
