% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_alevin}
\alias{read_alevin}
\title{Read in counts data processed with Alevin or Alevin-fry}
\usage{
read_alevin(
  quant_dir,
  mtx_format = FALSE,
  intron_mode = FALSE,
  usa_mode = FALSE,
  which_counts = c("spliced", "unspliced"),
  round_counts = TRUE,
  tech_version = NULL
)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.}

\item{mtx_format}{Logical indicating if input data is in matrix market format.
Default is FALSE.}

\item{intron_mode}{Logical indicating if the files included alignment to intronic regions.
Default is FALSE.}

\item{usa_mode}{Logical indicating if Alevin-fry was used, if USA mode was invoked.
Implies the input data is in matrix market format.
Default is FALSE.}

\item{which_counts}{Which type of counts should be included,
only counts aligned to spliced cDNA ("spliced") or all spliced and unspliced cDNA ("unspliced").
Applies if `intron_mode` or `usa_mode` is TRUE.
Default is "spliced".}

\item{round_counts}{Logical indicating in the count matrix should be rounded to integers on import.
Default is TRUE.}

\item{tech_version}{Technology or kit used to process library (i.e. 10Xv3, 10Xv3.1).}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix.
}
\description{
Read in counts data processed with Alevin or Alevin-fry
}
\examples{
\dontrun{

# Import output files processed with either Alevin or Alevin-fry with alignment to
# cDNA only, including only spliced cDNA in final counts matrix
read_alevin(quant_dir)

# Import output files processed with either Alevin or Alevin-fry with alignment to
# cDNA + introns and including all unspliced cDNA in final counts matrix
read_alevin(quant_dir,
            intron_mode = TRUE,
            which_counts = "unspliced")

# Import output files processed with alevin-fry USA mode
# including all unspliced cDNA in final counts matrix
read_alevin(quant_dir,
            usa_mode = TRUE,
            which_counts = "unspliced")

}
}
