% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellranger.R
\name{read_cellranger}
\alias{read_cellranger}
\title{Read in counts data processed with Cellranger}
\usage{
read_cellranger(quant_dir)
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}
}
\value{
SingleCellExperiment of gene x cell counts matrix
}
\description{
Read in counts data processed with Cellranger
}
