% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kallisto.R
\name{read_kallisto}
\alias{read_kallisto}
\title{Read in counts data processed with Kallisto}
\usage{
read_kallisto(quant_dir, intron_mode = FALSE, round_counts = TRUE)
}
\arguments{
\item{quant_dir}{Path to directory where output files are located.}

\item{intron_mode}{Logical indicating if the files included alignment to intronic regions.
Default is FALSE.}

\item{round_counts}{Logical indicating in the count matrix should be rounded to integers on import.
Default is TRUE.}
}
\value{
SingleCellExperiment of unfiltered gene x cell counts matrix
}
\description{
Read in counts data processed with Kallisto
}
\examples{
\dontrun{

# import output files processed with kallisto with alignment to cDNA + introns,
# including all unspliced cDNA counts in final counts matrix
read_kallisto(quant_dir,
              intron_mode = TRUE,
              which_counts = "unspliced")
}
}
