% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevin.R
\name{read_usa_mode}
\alias{read_usa_mode}
\title{Read in counts data processed with Alevin-fry in USA mode}
\usage{
read_usa_mode(quant_dir, which_counts = c("spliced", "unspliced"))
}
\arguments{
\item{quant_dir}{Full path to directory where output files are located.}

\item{which_counts}{If intron_mode is TRUE, which type of counts should be included,
only counts aligned to spliced cDNA ("spliced") or all spliced and unspliced cDNA ("unspliced").
Default is "spliced".}
}
\value{
unfiltered gene x cell counts matrix
}
\description{
Read in counts data processed with Alevin-fry in USA mode
}
