% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_anndata.R
\name{sce_to_anndata}
\alias{sce_to_anndata}
\title{Convert SingleCellExperiment objects to AnnData file stored as HDF5 file}
\usage{
sce_to_anndata(
  sce,
  anndata_file,
  x_assay_name = "counts",
  compression = "gzip",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object to be converted to AnnData as an HDF5 file}

\item{anndata_file}{Path to output AnnData file. Must end in `.h5`, `.hdf5`, or `.h5ad`}

\item{x_assay_name}{Name of assay in SCE object to save as X in AnnData. Default is "counts".}

\item{compression}{Type of compression to use when exporting hdf5 files. Options are
"none", "gzip", or "lzf". Default is "gzip".}

\item{...}{Any arguments to be passed into zellkonverter::writeH5AD or zellkonverter::SCE2AnnData}
}
\value{
original SingleCellExperiment object used as input (invisibly)
**Note that any columns present in the `rowData` of an SCE object that contains
duplicated information, e.g. duplicate gene identifiers, are converted to
categorical data by the `anndata` package.
}
\description{
Convert SingleCellExperiment objects to AnnData file stored as HDF5 file
}
\examples{
\dontrun{
sce_to_anndata(
  sce = sce_object,
  anndata_file = "test_anndata.h5"
)
}
}
