% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_anndata.R
\name{sce_to_anndata}
\alias{sce_to_anndata}
\title{Convert SingleCellExperiment objects to AnnData file stored as HDF5 file}
\usage{
sce_to_anndata(sce, anndata_file)
}
\arguments{
\item{sce}{SingleCellExperiment object to be converted to AnnData}

\item{anndata_file}{Name of}
}
\value{
Converted AnnData object
}
\description{
Convert SingleCellExperiment objects to AnnData file stored as HDF5 file
}
