% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_seurat.R
\name{sce_to_seurat}
\alias{sce_to_seurat}
\title{Convert SingleCellExperiment object to Seurat object}
\usage{
sce_to_seurat(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
Seurat object
}
\description{
Convert SingleCellExperiment object to Seurat object
}
\examples{
\dontrun{
sce_to_seurat(sce = sce_object)
}
}
