% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_seurat.R
\name{sce_to_seurat}
\alias{sce_to_seurat}
\title{Convert SingleCellExperiment object to Seurat object}
\usage{
sce_to_seurat(sce, assay_name = "counts")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{The assay name (default "counts") to include in
the Seurat object}
}
\value{
Seurat object
}
\description{
Convert SingleCellExperiment object to Seurat object
}
\examples{
\dontrun{
sce_to_seurat(sce = sce_object)
}
}
