% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_within_batch_ari.R
\name{within_batch_ari_from_pcs}
\alias{within_batch_ari_from_pcs}
\title{Calculate within-batch ARI using provided PCs for use in integration metric calculations}
\usage{
within_batch_ari_from_pcs(
  merged_sce,
  pc_name,
  batch_column = "library_id",
  BLUSPARAM = bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard")
)
}
\arguments{
\item{merged_sce}{The merged SCE object containing data from multiple batches}

\item{pc_name}{The name used to access the PCA results stored in the
SingleCellExperiment object}

\item{batch_column}{The variable in `merged_sce` indicating the grouping of interest.
Generally this is either batches or cell types. Default is "library_id".}

\item{BLUSPARAM}{A BlusterParam object specifying the clustering algorithm to
use and any additional clustering options. Default is
`bluster::NNGraphParam(cluster.fun = "louvain", type = "jaccard")`}
}
\value{
Tibble with three columns: `ari`, representing the calculated ARI values;
  `batch_id`, the batch id associated with each `ari`; `pc_name`, the name
  associated with the pc results
}
\description{
Calculate within-batch ARI using provided PCs for use in integration metric calculations
}
